Remove support for Bluefruit LE UART Friend
Header is removed for UART Friend.  Supporting only SPI Friend going forward.
Add a BLE_RESET signal to the SPI Friend BLE header.  Don't plan on using it, but will be nice if the header is used as generic expansion.
Move the BLE_PWR_EN signal to make room for the BLE_RESET signal.